FIX: add backwards compability for older browsers - Install `babel-cli` - Move source to `./src` directory + Update tests - Add `npm run build` command
