+++ test/utils.spec.js
@@ -1,4 +1,4 @@
-const utils = require('../lib/utils')
+const utils = require('../src/lib/utils')
 
 describe('buildRequestMethod', () => {
   let request
